Experimental Reactor support.
This adds support for a new experimental "Reactor" executable model.

The "Commands" and "Reactors" concepts are introduced here:
https://github.com/WebAssembly/WASI/issues/13

A companion Clang patch, which just consists of using the new
reactor-crt1.o and Reactor-specific entry point name, is here:
https://reviews.llvm.org/D62922

Instead of an entrypoint named "_start", which calls "main", which
then scopes the lifetime of the program, Reactors have a
"__wasi_unstable_reactor_start" function, which calls "reactor_setup".
When "reactor_setup" exits, the intention is that the program should
persist and be available for calling.

At present, the main anticipated use for this is in environments like
Node, where WASI-using modules can be imported and don't necessarily
want the semantics of a "main" function.

The "unstable" in "__wasi_unstable_reactor_start" reflects that this
Reactor concept is not yet stable, and likely to evolve.