More cleanups.
Make reactor_setup a weak external, and factor out common code between
the crt1.c implementations into a shared header.

This renames basics/include to basics/headers/public, which is more
consistent with the directory structures in libc-bottom-half and
libc-top-half, and which allows us to have basics/headers/private, for
the new crt1.h.